% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifunction.R
\name{uifunction}
\alias{uifunction}
\title{Ui for shiny}
\usage{
uifunction(n, itermax, x1mu, x2mu, varx1, varx2, covx1x2)
}
\arguments{
\item{covx1x2}{}
}
\value{
input and output for server
}
\description{
Ui for shiny
}
\examples{
ui = uifunction(300,7000,0,0,200,500,50)
}
