% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbarthetadist.R
\name{xbarthetadist}
\alias{xbarthetadist}
\title{xbarthetadist function}
\usage{
xbarthetadist(n, iter, mu, sigma)
}
\arguments{
\item{sigma}{}
}
\value{
a matrix composed of iterated results on four categories including xbar1, xbar2, thetae, and thetar
}
\description{
This is a constuctor function that is used to create thetaeigen and thetarot from sample variance covariance matrix
}
\examples{
 mu1=c(0,0); sigma1 = matrix(c(100,40,40,400) nr=2,nc=2,byrow=FALSE); xbarthetadist(n=30,iter=1000,mu=mu1,sigma=sigma1)
}
